hppa: Allow working with larger code-constant vectors
  * Bump the limit from (signed-byte 11) (about 250 constants,
give or take) to effectively unlimited, by using LIP (which is
never live during LOAD-CONSTANT) as a temporary.

  * LIP won't pair during GC when it holds one of these offsets
unless it reaches the start of heap space, which is
currently #x4b000000, larger than all of the heap spaces put
together, let alone any one heap space individually.  It's also
larger than the 26 bits of offset that the code header allows to
describe the number of boxed words in a code object (24 bits of
header data, plus word shift).