hppa: Fix sign computation for TRUNCATE quotient
  * Under some combination of circumstances, which I'm having
trouble precisely articulating but includes (truncate -4 -4), the
quotient returned would have the incorrect sign (in this case, it
would be -1, and should be 1).

  * The root cause turned out to be an instruction that does a
conditional nullify as part of a remainder-correction step getting
executed when no remainder-correction was needed, thus nullifying
the test to see if the quotient sign needed correction.

  * Fix, by marking the test for needing the remainder-correction
as nullifying the following instruction if its branch is taken.