manual: Fix treatment of non-ASCII characters in the texinfo files.
There is only one currently, in the section about Unicode support;
it was broken in the HTML output and invisible in the PDF output.
To fix this, declare the correct document encoding.

Besides fixing these two output formats this also makes makeinfo add
"Local Variables" footers to the info files it generates, declaring
their character encoding, which is supposed to help info readers to
display such characters correctly.