Ensure that arbitrary DEFUN clears struct slot accessor transforms.
Excluding the DEFUNs that define out-of-line accessors, of course.

This is done in a way that eliminates reliance on the evaluator to expose
a function's sexpr to decide whether to remove a source-transform,
and more generally solves the problem cited in lp#540063 that compiling
inline/notinline calls were at odds with one another after modifying the
definition of a function that was auto-generated by DEFSTRUCT.

And don't ever store inline expansions for slot accessors.