optimize ORDERED-RANGES-MEMBER
(safety 0) isn't generally a good idea, but in this case it's a
substantial speed saving in a local body of code.  In an ideal world the
compiler would be able to prove that there are no unsafe array accesses
and that everything has appropriate times, but in the absence of such a
sufficiently smart compiler, tell it that we know what we're doing.