improve SB-UNICODE:GRAPHEMES
Build graphemes out of new mapper functions and adjustable arrays.  The
lower-level mapping interface is probably not quite right yet (it's hard
to see how e.g. a regex engine could use it at the moment, and it might
need :START and :END arguments.) but it's a stab at making the
investigation of grapheme contents of strings accessible without consing
up a ridiculous amount of stuff.