Revamp disassembler's instruction-defining macros.
All uses of the baffling VALSRC structures are eliminated,
many uses of (EVAL-WHEN (:COMPILE-TOPLEVEL) ...) become unnecessary,
and the code-folding optimization is far simpler.

Compile-time definitions for :PRINTER, :PREFILTER, and :LABELLER options
are never needed. The helper functions resulting from FIND-PRINTER-FUN etc
are not compiled during cross-compilation, but instead just-in-time when
DISASSEMBLE is invoked by the user. Disassembly of any given instruction
format has a slight delay the first time it's encountered, but it's not
bad, and if anyone cared, precompilation could be easily be done.
[One suggestion was that compilation not be done at all, but then we'd
need to make an interpreter for abstract printer specs, or else interpret
a subset of Lisp code resulting therefrom without relying on EVAL]