Fix result sequence creation in SEQUENCE:MAP
It passed (class-of result-prototype) to MAKE-SEQUENCE which, in the
best case, would go back to the prototype of the class and call
SEQUENCE:MAKE-SEQUENCE-LIKE on it.

All subtype queries and whatnot performed by MAKE-SEQUENCE before
finally calling SEQUENCE:MAKE-SEQUENCE-LIKE are unnecessary in the
common case of SEQUENCE:MAP being called from MAP in which case the
result type is an extended sequence.

Instead, directly call SEQUENCE:MAKE-SEQUENCE-LIKE on RESULT-PROTOTYPE.

This is what SEQUENCE:{CONCATENATE,MERGE} already do.