Declare interpreter env constructors inline after the defstruct.
This avoids the inline expansion capturing its macro environment,
which stupidly included the whole macrolet form itself.
The compiler can reconstitute a defstruct constructor without using
a stored inline expansion. Probably we should never store an expansion
even when the user "wanted to", because the implementation of inlining
is supposed to be a black box.