hppa: "Proper" read-only-tramps support
  * The actually-used versions of the funcallable-instance and
closure tramps were already in read-only space, but there was
still the copy of the closure tramp in hppa-assem.S in order to
keep genesis from pitching a fit...  And the undefined tramp was
ONLY in hppa-assem.S, preventing it from working properly on HPUX.

  * So, stick everything in assembly/hppa/tramps.lisp, remove the
hppa-assem.S copy of those things which were in hppa-assem.S, fix
up the various references in the compiler backend, add the feature
to make-config.sh...

  * ... and, while we're here, try (and fail) to address two test
failures for debug.impure.lisp / (BACKTRACE UNDEFINED-FUNCTION
BUG-xxx).  Either something isn't quite right with the new tramp
logic or there's something else confounding the backtrace, but the
situation is no worse and arguably better, so we'll run with this
for now.