Reinstate "Change layout bitmap so that 0 = raw, 1 = tagged"
The problems on 32-bit were:
 - fixnum_value() claims to be able to return a negative number,
   however it performs a logical right shift instead of arithmetic
   shift, so will do the wrong thing for negatives. This is probably a bug.
 - a couple places were missing a use of make_fixnum()

This change worked by accident on 64-bit because omission of make_fixnum
in scav_instance simply meant that the bypass of instance_scan_interleaved
never happened, which was ok because the general case worked.
The general case worked only because there are no builtin structures with
enough slots such that it mattered whether the bitmap was considered signed
or unsigned. i.e. even if it should look like an arbitrary number of tagged
slots due to the sign bit of 1, indexing never got that high.

I'm still suprised that no test failed though.