Optimize xref storage for the most frequently referenced names
Based on the observation that a relatively small number of names appear
in the majority of all cross references, it makes sense to add a
special, more compact encoding for these names.

To this end, names stored in the global vector
**MOST-COMMON-XREF-NAMES-BY-INDEX** are represented as indices into that
vector and are never stored in individual xref data vectors. Reserving a
part of the encoding space (currently 8-bit by default) for the most
frequent names makes the representations of all other names less compact
but still saves lots of space due to the vastly different occurrence
frequencies.

This commit only adds the machinery,
but **MOST-COMMON-XREF-NAMES-BY-{INDEX,NAME}** are not populated yet.