Fix parallel-fasl-load crashing.
In rare cases, the fun and raw_addr slots of an fdefn can refer to two
different simple-funs because set-fdefn-fun is not atomic.
When that happened, GC would not fixup the raw addr (nor even preserve
the referenced code component), despite it being possible for a thread
to compute its address implicitly, such as via "JMP [RAX+9]" on x86-64.

Rather than try to make the setter atomic - which might be impossible -
treat the raw_addr as a pointer that is not necessarily related to the
tagged function, and ensure that it survives GC if the fdefn does.