Actually place static functions first in immobile space.
And make the default roots for immobile space DFO meaningful,
but, more importantly, fix a potential crash.

SAVE-LISP-AND-DIE always passes its :ROOT-STRUCTURES argument
into CHOOSE-CODE-COMPONENT-ORDER, so the apparent default in the latter
had no effect, because S-L-A-D defaults the roots to NIL.
That's fine in and of itself; however if you *did* supply roots,
then they would precede the static functions, which could cause
a subsequent S-L-A-D to produce a new DFO for the roots - if changed -
which could consume different amounts of immobile space, which could
move static functions from where they previously were, which is bad.

Also annotate a flaw in 'save.lisp'