Ctors handle refedinition of constants used as initarg values better
Previously, keys for cached constructors included names of symbolic
constants when those were used as values of initargs. As a result,
recompiling such a constructor after redefining the corresponding
constant would use the cached constructor since the key did not
reflect the change, thus not picking up the new value.

Now, the values of symbolic constants contribute to the keys instead
of their names.

fixes lp#1644944