Remove obsolete comment, added in change 4c5a011c
OUTPUT-NUMBER never existed and likely never will.

Note also that OUTPUT-FOO functions are in general authoritative for ugly [sic]
output only. In contrast, OUTPUT-OBJECT *decides* whether to print prettily.
If for whatever reason you had a dispatch entry for BIGNUM, it's probably wrong
to call a specialized numeric printer in lieu of the pretty method if pretty=T.