x86-64: enable static linkage for all immobile-space code
Static linkage means that instructions are encoded as
"CALL #<FUNCTION FOO>" instead of "CALL #<FDEFN FOO>".
This is the new default if #+immobile-code is enabled.
There is no change to language semantics - if you redefine
a function F that was statically linked, the static links
to F are removed and replaced with dynamic calls, and they
stay that way unless the core is saved to disk again.

If people don't like the extra CPU cycles associated with
redefinition of functions dumped in a core, we will need to
export and document STATICALLY-LINK-CORE from SB-EXT,
so they users can invoke as they see fit. For the time being,
it's one less thing to explain, other than "it's faster"