Scrub cold-init-only symbols better.
The macroexpansion of WITH-SINGLE-PACKAGE-LOCKED-ERROR
contained a use of /NOSHOW0 which, by the time "pcl/defcombin"
is compiled, got uninterned. This is ugly, so delete the use.

And we weren't calling FMAKUNBOUND on symbols that got uninterned
by the ad-hoc code provided per-package (not the symbols whose
names start with "!" and are always uninterned) so could retain
code accidentally if we perchance didn't GC the symbol. Fix that.

And move all the uninterning logic out of cold-init.