Shorten funcallable-instance by 2 words if #!+compact-instance-header
Same as for non-funcallable instances, put the hash code in the header
word of the slot vector, delete a primitive object slot, and its padding.
This is mainly for consistency, not performance; it would be surprising
if the amount of memory consumed by GF objects mattered to anyone.
However, all other things being equal, shorter is better.