Conventionalize style in purify.c and make it more robust
Never compute an object size "by hand" - just call the sizer.
While this adds overhead - and we avoid that in gc-common, instead
choosing to repeat the size calculation in each transport function -
it is not a win to do that when Lisp is about to exit anyway.
This is a reduction in maintenance cost. No behavior change intended.