Delete some hubbub and insanity.
With all the commentary about how complicated it is that %SIMPLE-FUN-SELF
is defined differently for x86oids, you'd think it would actually be relevant
to non-x86oids. But it isn't. Kill that nonsense, and never directly call
%SIMPLE-FUN-SELF. As in not ever, unless via %CLOSURE-FUN.
And obviously do not call it on the return value of %FUN-FUN,
which is guaranteed to be a SIMPLE-FUN with absolute certainty.