Compile out some GC assertions
- the scav_x_pointer family of functions are only invoked via
  widetag dispatch; low bits are necessarily correct.
  Similarly trans_x functions are invoked on pointers in from_space.

- trans_list and friends allocate memory, so it's not possible
  for them to have "copy == object" unless they accidentally return
  the original object, which they don't.

While these are potentially helpful for debugging, they don't make
a lot of sense once programming errors have been eliminated.

Using the same test approach as in the prior patch, but with revision
d84042ff as the reference baseline, the net effect of this and all
intervening changes was -6.5% GC time averaged over 100 trials.