Better sharing of symbol print-names.
Do more thorough collapsing because:
"X3J13 voted in March 1989 to specify that it is an error
to modify a string being used as the print name of a symbol."
and remove logic from Genesis that did the same thing.

Also prefer to use base strings in the common pattern of
(INTERN (FORMAT NIL "~A-~A")). While it would be nicer to have INTERN
always pick the minimal space representation no matter what,
it seems like some people think that (INTERN "FOO") - though permitted
to copy "FOO" - must copy to the same type string, which is unspecified
as far as I can tell. That said, it's unarguable that (FORMAT NIL ...)
produces any particular type of string.
And this saved us >7MB in our application, so ... there you go.