2-for-1: fix hopscotch tests and cheneygc
Initializing the cache of memory mappings by asking for a fixed amount
and considering it divisible in half does not do the right thing
if the size of each isn't a multiple of getpagesize(). When the first half
was unmapped, the kernel rounded to its page size, effectively unmapping both.