Change how pathname namestrings are memoized.
Instead of memoizing #'NAMESTRING, build up a table of interned pathnames,
and memoize the namestring in a slot of the pathname object.
The pathname cache is cleared out in SAVE-LISP-AND-DIE,
or at other times as decided by the memoization algorithm.

And improve pathname hashing so that it's actually useful for
directory names having more than three or four subparts to them.