Improve SXHASH on known keywords.
There's no need to test for whether the symbol-hash is nonzero.

We've had discussions as to whether compile-time keywords are guaranteed
to be keywords at runtime (lp#922952 which actually claimed "if and only if"
which may not be true), but I see no way that this transform can be wrong.
First of all, the compiler seldom, if ever, deduces that a variable is of type
KEYWORD; thus the transform should only affect forms such as
(SXHASH (THE KEYWORD MYSYM)). Second, only if a compile-time keyword is
a non-keyword at runtime could the code go awry; but that's a case of misleading
the compiler, which is undefined behavior. And finally, uninterning a keyword
at random (which is disallowed, if I understand correctly) doesn't affect
correctness of the code - all interned symbols have hashes, so if a symbol was
ever a keyword (i.e. interned), then it has a nonzero hash value.