Fix accidental pessimization of BOA constructor
If a slot appeared in any BOA ctor as an &AUX binding
with or without an initial value, it was deemed unsafe.
It should only be unsafe if there is no initial value.

Also add some remarks about how to further optimize the case
of BOA constructors that leave slots uninitialized.