Use similar is_valid_lisp_address() code for everybody
It was apparently the intent - who knows when - to ask the OS whether a
Lisp heap address was valid. None of the current variations do that,
not counting osf1, which had a syscall for it.  So move the linux variant
to os-common, deleting the others except win32 which needs this function within
the fault handler and acquires a mutex around the per-thread range tests.