Store unbound-marker, not 0, in uninitialized structure slots.
This is not intended to be a full rendering of the CLOS unbound slot notion.
It is only designed to trap slot reads on structure instances made by
BOA constructors with &AUX variables that don't specify initial values.
In particular there is no SLOT-UNBOUND invocation,but only a TYPE-ERROR,
the same as before. Aside from being efficient, this avoids an edge case
in which 0 was a legal value for the slot, wherein slots initialized
by accident could not be distinguished from those specified as uninitialized.

There's no reason this same change couldn't be done for structures
whose storage type is (VECTOR T) or LIST.