Improve ATOMIC-POP in a few ways
(1) Don't reload from place on a failed CAS; use the CAS result.
It appears that the previous code tried to do this, but the author
mistook the behavior of "LOOP FOR" for "LOOP WITH".

(2) Elide the final LISTP test. The cell whose CDR was stored back
into place already satisfied LISTP. CAR doesn't need another test.

(3) The simple form of LOOP generates significantly better code here,
including elision of some moves and a useless initialization.