Dynamic space relocation, part 1 of 2
When automatic relocation is in effect, a saved core's dynamic space
start is taken as a hint, not a requirement. If not mapped at the preferred
address on image startup, all spaces will be scanned and pointers into
dynamic space adjusted by the difference between desired and actual address.

This should enable running under sanitizing interceptors such as msan
without having to rebuild with a different dynamic space start.