x86-64: Fix insanely subtle bug in get_pseudo_atomic_atomic()
Pseudo-atomic is true when thread->pseudo_atomic_bits is nonzero in any bit
except the least significant. (The LSB indicates that a signal was deferred
during the pseudo-atomic instruction sequence)

Change https://sourceforge.net/p/sbcl/sbcl/ci/4c400df29038a283 decided cleverly
to use the value in RBP rather than an immediate operand on a MOV instruction.
I'm not sure of the veracity of the speedup claim, but sure, RBP is nonzero.

However, take a very close look at the declared type of get_p_a_a(), and see
that truncation to 32 bits has to occur on (X & (~1)). Therefore if X has 32
trailing zeros, the truncated value is 0. This triggered a failure in alloc()
when RBP contained 0x7fff00000000 or any value similarly aligned.

Forcing a single bit result via a '!= 0' comparison corrects the problem.