Allow building C runtime without SBCL_PREFIX set
This is a very sane thing to do, as it avoids wiring in pathnames,
and allows copying an install tree anywhere. (People often don't realize
that "cp -r old new" will have the "new" binary use the "old" core)

We should probably make this the default, and make it effective whenever
SBCL_HOME is not set. However, strictly speaking, that change is incompatible.

So this feature is advertised with only a small mention in NEWS rather than
fanfare and revision of the INSTALL guide, for now. Those who have been
clamoring for this mode of operation will figure it out.

See also https://bugs.launchpad.net/sbcl/+bug/666086