Fix heap relocation, for real now
The test "layout > (lispobj)where && needs_adjusting(layout)" contained
two flaws, one on each side of the conjunction, naturally.

One flaw was that the address of the layout *after* adjusting should be
compared to the address of the instance to see whether the slots of layout
were already adjusted.

The other was that needs_adjusting() is irrelevant, but did no harm at present,
only because immobile space doesn't get relocated. To see this:
* if the layout is in dynamic space, then since dynamic space is moving,
  the layout is moving, therefore "needs_adjusting" is tautologically 1
* if the layout is in immobile space, then needs_adjusting is 0,
  but because immobile space was fixed in its entirety before handling
  dynamic space, the layout's bitmap, if a bignum, was already adjusted

"core.test.sh" with the mmap injection stub is still not enough
to prove that this change actually fixed anthing, but couldn't hurt.