Fix #+(and (not sb-thread) sb-dynamic-core) builds
This feature combination was broken ever since dynamic-core was implemented.
Incredibly, and truly so, the x86-64 build _seemed_ to run, but the broken-ness
was obvious, as "boxed_region" was present in linkage space twice: once as a
data reference used by ADDRESS-BASED-COUNTER-VAL, and once as a function.

So how could it work? The linkage entry for "boxed_region" as a function
contained a word than when misinterpreted as an address, was too high to be
a pointer into the alleged allocation region:

* (disassemble 'cons)
...
; 0F7:       4C8B1C2510002020 MOV R11, [#x20200010]           ; boxed_region
; 0FF:       498D5310         LEA RDX, [R11+16]
; 103:       483B142518002020 CMP RDX, [#x20200018]           ; boxed_region

but [#x20200010] holds a JMP opcode:
* (sb-disassem::disassemble-memory #x20200010 8)
; Size: 8 bytes. Origin: #x20200010
; 0:       FF2500000000     JMP QWORD PTR [RIP]               ; boxed_region
; 6:       E083             LOOPNZ #x201FFF9B
the value of which:
* (format t "~X" (sb-sys:sap-ref-word (sb-sys:int-sap #x20200010) 0))
83E00000000025FF
is larger than the value in [#x20200018], the alleged end of boxed_region:
* (format t "~16,'0X" (sb-sys:sap-ref-word (sb-sys:int-sap #x20200018) 0))
0090000000000064

So the test for whether "free_ptr < end" fails, and alloc_tramp is invoked.
Except that on x86, the linkage values didn't conspire to make failure happen,
and inline allocation tried to proceed as if it worked, then crashed horribly.

A proper fix for this is not really worth doing, and the initial commit
of #+sb-dynamic-core (https://sourceforge.net/p/sbcl/sbcl/ci/f057566fe993f008)
said as much. But it also claimed that the feature combination was "disabled"
which I would have interpreted to mean that it actively checked for and
prevented that via the conditions in src/cold/shared.lisp