Impart consistency to accessing thread-local Lisp vars from C
The most frequent way to read a symbol value was with SymbolValue(), but
safepoint code usually opted for TlSymbolValue(). This left one to wonder
why/when SymbolValue was permitted to fallback to reading symbol->value
with multi-threads. The answer is: NEVER. The road less traveled was correct.
Additionally there was a minor annoyance of spelling; the setter for
SymbolTlValue was SetTlSymbolValue, not SetSymbolTlValue.

To reconcile these problems and improve the situation, there are two new
preprocessor macros: read_TLS() and write_TLS() each of which does exactly
what its name says. By design, the macros will fail to compile if a TLS
index was not assigned at compile-time.

In unithread builds, the macros simply reference 'symbol->value'.
Given that the TLS isn't used, it shouldn't be too much to suppose
that that's what they would do.