x86-64: Indirect alloc_tramp through immobile space
(Only if immobile space exists)
There is no extra indirection imposed - the caller jumps to the immobile
trampoline which jumps to the assembly code, not to the linkage-table.
Hopefully this allows immobile space to be mapped anywhere,
though there may be some more issues to track down.
Also, it works with or without dynamic-core in *features*.