Define weak hash table entry liveness tests in a '.inc' file
or "template metaprogramming in plain C for fun and profit"

The problem this deals with is as follows: To implement a different GC strategy
in the general framework provided by 'gc-common', you need to customize
from_space_p(), which is is critical to performance of pointer_survived_gc_yet()
which is needed by the 4 different hash-table entry liveness tests.
It would be inefficient if from_space_p() had to perform an internal switch
based on global state, or call out to something else. So this change pushes
the from_space_p() implementation into the bodies of the hash-table liveness
tests, allowing definition of different arrays of optimal functions.

Other changes:
- provide scav_weak_hash_tables() and scav_hash_table_entries()
  the array of tests and a function to call on each live k/v pair.

- augment scav_binding_stack() to accept a function to call

- move part of scav_weak_pointer() into a utility function

- flip tests in conservative_root_p() to make it more amenable to pending
  changes that disregard the generation. Better to fail on the address.