x86-64: fix bug in %SET-FDEFN-FUN
The code assumed that dynamic space would never be mapped at a sub-4GB address,
and that any funcallable-instance at a high address needed simple-izing, and
any at a low address didn't, due to a bug that compounded the address bug by
testing layout-bitmap incorrectly. The correct logic is that if the instance
is moveable by GC, or has only tagged slots, it needs a trampoline.