Delay storing simple-fun layout until code is loaded.
This will - after immobile space relocation works - allow loading fasls
into an image having a different immobile-space-start address from the
image that wrote the fasls. It's a separate concern from whether
coreparse and GC know to fixup layout in headers, which they do,
but not if the layout starts out completely wrong.