x86-64: Enliven funcallable-instances from unboxed words
This patch makes immobile_space_preserve_pointer() pin a funcallable-instance
if a stack word (REG_RIP from ucontext) points to it.

It is future-proofing against a sufficiently smart compiler and implementation
of PCL that sucks less. PCL holds on to GFs from a bunch of hash-tables,
so practically speaking, immobile-space GFs are all *permanently* live anyway.

But GC should not be correct due to the failings of others.
It should be correct period. In addition, we can't prove that people
aren't seeing this problem (maybe it actually happens with nonstandard GFs)