Make verify_gc() callable from Lisp at more-or-less any time
Specifically, this works quite well in single-threaded code:
  (alien-funcall (extern-alien "verify_gc" (function void int)) 1)

We just assume that pointers to pages with open allocation regions are
probably OK. There's no switch to enable or disable that degree of freedom-
If a page has an open region, it should always be allowed. Whether pages
erroneously have regions is a separate question, asserted elsewhere.

Additionally, if the VERIFY_AGGRESSIVE flag is specified, do a slow
and thorough check (as before) for pointer targets; but if not specified,
do a quick sanity check for a good header or cons, as opposed to no check
aside from the page bounds / forwarding-pointer / immobile-space hole.