Modernize C style of and deobfuscate gc_alloc_update_page_tables()
Once per decade or so it's a good idea to edit comments for correctness.

Witness pearls of wisdom like "the rest of the pages should be free"
to mean "... be *USED*" and "set up the gen, and scan_start_offset"
despite that those things were done when the region was newly opened.