Allow 1-page "large" vectors in scavenge_generations()
We don't now see any vectors taking up a single large-object page,
but it's a perfectly reasonable thing. We can teach the allocator
that a vector not meeting the large object size is a candidate for
being by itself if it uses up almost all of its page or pages.
If done, the loop invariant of !page_ends_contiguous_block_p()
needs to test that before assuming that a second page exists.