Perform less work for :KEY-AND-VALUE hash-table weakness.
Such tables do not need to be repeatedly visited in newspace scavenge.
If only one of the key or value per entry is live, nothing happens.
If both are known live, then nothing need happen, because no new things
become live by dint of visiting a live entry. So only one pass over the
table is required, and as late as possible so that nothing else can
become live. Do that pass just before smashing the non-surviving pairs.