Allow nonexistent "output/dynamic-space-size.txt"
It's main use is to inject a preprocessor #define for C.
This can be done in other ways.  While we also verify that the specified
value does not cause spaces collision at build-time, the default suffices
for that test, since ultimately the runtime value is what matters.

And rename :DEFAULT-DYNAMIC-SPACE-SIZE to :DYNAMIC-SPACE-SIZE
in !GENCGC-SPACE-SETUP.  It doesn't make any sense that one of the ways
to specify the same thing is named "default" when they're all equally defaulty.
(the build-time flag isn't --default, nor is the text file named that)