More tweaks to allow not-in-source-tree hermetic build.
One may now host compile+load all the files of the cross-compiler, then
relatively immediately (SETQ *TARGET-OBJ-PREFIX* #P"someplace") and
then (WITH-COMPILATION-UNIT (LOAD "src/cold/compile-cold-sbcl.lisp"))
without much else. You have to run tools-for-build/ucd.lisp first, but that
has no dependencies on target features or anything. make-host-2 is relegated
to post-compilation steps that are not strictly needed to make cold-sbcl.core

As a funny side-effect of this patch, the mysterious (and now gone)
*FLAME-ON-NECESSARILY-UNDEFINED-THING* became clearer- we still don't need it.
Prior to WITH-COMPILATION-UNIT surrounding execution of the xc, it emitted
thousands of warnings for undefined functions, among those, GET-SETF-EXPANSION.
Fine, except that we know that it is not legally permitted to define it in
conformant code. W-C-U defers the warning - so it's as if the right hand didn't
talk to the left: we still know that it can't be defined, but when it comes
time, we do it anyway and remove the deferred warning. This was a full WARNING,
not a style-warning, so it would have been fatal had it not been removed.
But the extra hack that said, "well, don't die", was just grotesque.