win32: Better processing of :HOME in UNPARSE-NATIVE-WIN32-NAMESTRING
For the (:HOME USERNAME) case, the code had a comment arguing not to
use USER-HOMEDIR-NAMESTRING and then proceeded to call

  (USER-HOMEDIR-PATHNAME USERNAME)

which is clearly wrong since the optional parameter of
USER-HOMEDIR-PATHNAME is a host, not a username.

So we call USER-HOMEDIR-NAMESTRING which actually accepts a username
and try our best to make sure there are no conflicting device
components.