x86-64: Fix ALLOCATE-VECTOR-ON-STACK for #+msan, again
This was almost a rehash of an old bug which was the reason for arg TNs having
crazy lifetimes. Rev 1785d1e18c was the last to touch the lifetimes, saying
"If WORDS and LENGTH were packed in the same TN, all manner of badness arose"
Unfortunately the msan code succumbed to that exact problem.

Prior to the aforementioned change, there was a bug that (MOVE ECX WORDS)
before the ALLOCATION would clobber LENGTH if it and WORDS were both in ECX,
because LENGTH was not declared to conflict with ECX - LENGTH's life allegedly
ended earlier. The fix claimed to "Adjust lifetimes to make sure it [double-use
of a TN] doesn't happen," but it didn't do that, and I can't see why its author
thought it would. Instead it did something good enough, which was to extend
LENGTH's life up to :eval which does conflict with ECX. But at some point later,
we rearranged code to avoid clobbering ECX until *after* the ALLOCATION call.

The time around, it was a problem if RAX held both LENGTH and WORDS, so:
(1) Nikodemus's change clearly did not do what it said.
(2) Developers have a hard time perceiving the extremely subtle freedom
that the register packer affords itself.

So having fixed the RAX bug, I removed the non-default arg TN lifetimes which
only muddy the water. 'result' can not be packed in the same location as any
input, so computing it first and storing type/length works just great.
'result' can also act as a temp since it's redundant with rsp inside the vop.