Fix type checking of unused mv-call &rest args.
An mv-call is transformed to mv-bind when the lambda list consists
only of &optional and an unused &rest, declaring the type of the &rest
parameter will make it used, but it's checked in one of the dispatch
functions and will not show up in LEAF-REFs.
If the type is declared and conflicts with LIST don't convert such a
call to mv-bind.