Fix consing inside foreign thread callbacks.
Setting *gc-inhibit* to nil is not the proper way to enable gcing, the
pending interrupts need to be checked.
GCing initiated by the new thread needs to be delayed only because
*CURRENT-THREAD* is not yet initialized and the locks grabbed by
SUB-GC do not function without it.
Simplify MAKE-THREAD and INITIAL-THREAD-FUNCTION-TRAMPOLINE using the
fact that the new thread can cons freely after *current-thread* is
set.
And in enter-foreign-callback, wrap consing in WITHOUT-GCING until
*CURRENT-THREAD* is set.