safepoint: Correctly balance CSP manipulation in safepoint trap handler
  * Safepoint trap handlers invoked from lisp code need to mark
the thread executing them as being "in alien" code before doing
anything that can end up waiting on a condition variable.  And
then need to restore their status as being "in lisp" code before
returning.

  * Fix one place where the status wasn't being restored correctly.
This didn't break the GC cycle that was being executed, but it
"poisoned the well" such that there was a chance that the thread
would be thought to be "in alien" while it was actually "in lisp",
which can lead to a deadlock due to the gc_state.phase_wait[]
counts being wrong.